Move the now fast tests out of integration tests
